Add support for regex matching
Note: This is currently limited and needs to support escaped values, and
a way to pass flags to the re.match function. Without this there is not
way to search for \s (any whitespace), or case insenstive searches.

I don't know the antlr4 syntax well enough to exented the parser to
include these in the lexer.